+++ src/InteractiveSeven.Core/Memory/IEquipmentAccessor.cs
@@ -5,8 +5,8 @@ namespace InteractiveSeven.Core.Memory
     public interface IEquipmentAccessor
     {
         void SetCharacterWeapon(CharNames charName, byte weapon);
-        void SetCharacterArmlet(CharNames charName, int armlet);
-        void SetCharacterAccessory(CharNames charName, int accessory);
+        void SetCharacterArmlet(CharNames charName, ushort armlet);
+        void SetCharacterAccessory(CharNames charName, byte accessory);
         byte GetCharacterWeapon(CharNames charName);
         byte GetCharacterArmlet(CharNames charName);
         byte GetCharacterAccessory(CharNames charName);
